hive: amd64+mu: Add timer management + TSC support
When the OpenModality kernel boots and it is time for clock devices to
be initialized. The kernel calls into the CORE layer's timer subsystem
which creates a knode directory to store clock devices before calling
into the machine unifier layer for all clock devices to be initialized
and registered to that knode directory. Later on, it may be possible to
create drivers that perform high-level device specific operations while
using an interface decoupled from machine specifics.

Signed-off-by: Ian Moffett <ian@osmora.org>